Reconfigured the build.gradle files in subprojects demo and utils

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -3,6 +3,6 @@ package com.epam.utils;
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
          //here magic will happen
-        return Integer.parseInt(str) > 0;
+        return Double.parseDouble(str) > 0;
     }
 }
